Basic Distributions

These distributions behave the same way as the ones from Distributions.jl. However, as described in KernelDistributions.jl, scalars are always sampled on the CPU even if a CUDA.RNG is provided. They are named accordingly with Kernel<distribution name> prefaced.

KernelDistributions.BinaryMixture — Type

BinaryMixture(dist_1, dist_2, weight_1, weight_2)

Mixture model of two distributions optimized for calculations in the logarithmic domain. Weights are automatically normalized and transformed to log domain in inner constructor.

KernelDistributions.KernelDirac — Type

KernelDirac(value)

Type stable implementation of a Dirac distribution which has all it's mass at the value

KernelDistributions.KernelExponential — Type

KernelExponential(β)

Type stable implementation of an (negative) Exponential distribution. Uses the scale parameter β so the pdf is defined as: inv(β)*exp(inv(β)).

KernelDistributions.KernelNormal — Type

KernelNormal(μ, σ)

Type stable implementation of a Normal distribution. Parameterized by the mean μ and standard deviation σ.

KernelDistributions.KernelUniform — Type

KernelUniform(min, max)

Type stable implementation of a Uniform distribution. Parameterized by the support [min,max]. If the value is outside the support, logdensityof returns -Inf (alternative TailUniform).