Special Distributions

Differ from the behavior of Distributions.jl.

KernelDistributions.Circular

KernelDistributions.CircularUniform

KernelDistributions.SmoothExponential

KernelDistributions.TailUniform

Bijectors.logabsdetjac

Bijectors.transform

Bijectors.transform

KernelDistributions.Circular — Type

Circular

Transform ℝ → [0,2π)

KernelDistributions.CircularUniform — Type

CircularUniform

Generates uniform random numbers ∈ [0,2π]. If the lower/upper bound is exceeded the value continues at the other bound (2π+0.1=0.1).

Bijectors.logabsdetjac — Method

logabsdetjac(Circular, x)

mod2pi will not be zero for n2π, thus the discontinuity will not be reached. Thus, the log Jacobian is always 0.

Bijectors.transform — Method

transform(Circular, y)

Uses mod2pi to transform ℝ to [0,2π].

Bijectors.transform — Method

transform(Circular, x)

Transform from [0,2π] to ℝ. In theory inverse of mod does not exist, in practice the same value is returned, since [0,2π] ∈ ℝ

KernelDistributions.SmoothExponential — Type

SmoothExponential(min, max, β, σ)

Smooth truncated exponential distribution by convolving the exponential with a normal distribution: Smooth = Exp ⋆ Normal This results in smooth min and max limits and a definition on ℝ instead of ℝ⁺

Does not support truncated of Distributions.jl since it is a smooth truncation of the exponential distribution.

KernelDistributions.TailUniform — Type

TailUniform(min, max)

Acts like a uniform distribution of [min,max] but ignores outliers and always returns 1/(max-min) as probability.